+++ test/runtests.jl
@@ -12,5 +12,6 @@ using Aqua
     include("test_codes.jl")
     include("test_voltlvl.jl")
     include("test_io.jl")
+    include("test_extract.jl")
     include("test_grid.jl")
 end